+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
-- 
+- Added allow higher skill level config
 ### Fixed:
 - 
 ### Changed:
-- Removed libgui
+-